+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -11,6 +11,7 @@ Latest (add to next release)
 ### Changed
 - Restructured component object into page object model
 - Refactored component and accessibility tests
+- Extracted svgTransform to global test utils for reference by other packages
 
 
 v0.58.0